Separate out ToggleSegmentedControl to reduce duplication
