+++ src/lib/features/settings/ToggleItem.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
   import type { ToggleConfig } from '$lib/types/index';
-  import IconEye from '$lib/app/icons/IconEye.svelte';
-  import IconEyeSlash from '$lib/app/icons/IconEyeSlash.svelte';
+  import ToggleSegmentedControl from '$features/toggles/components/ToggleSegmentedControl.svelte';
 
   interface Props {
     toggle: ToggleConfig;
@@ -10,8 +9,6 @@
   }
 
   let { toggle, value = $bindable('show'), onchange = () => {} }: Props = $props();
-
-  const icons: Record<'hide' | 'show', typeof IconEye> = { hide: IconEyeSlash, show: IconEye };
 </script>
 
 <div class="card">
@@ -22,25 +19,10 @@
         <p class="description">{toggle.description}</p>
       {/if}
     </div>
-    <div class="segmented" role="group" aria-label="Visibility">
-      {#each (['hide', 'peek', 'show'] as const) as option (option)}
-        <button
-          type="button"
-          class="segment-btn {value === option ? 'active' : ''}"
-          onclick={() => { value = option; onchange({ toggleId: toggle.toggleId, value: option }); }}
-          aria-pressed={value === option}
-          title={option.charAt(0).toUpperCase() + option.slice(1)}
-        >
-          {#if (option === 'hide' || option === 'show')}
-            {@const IconComponent = icons[option]}
-            <span class="segment-icon">
-              <IconComponent />
-            </span>
-          {/if}
-          <span class="segment-label">{option.charAt(0).toUpperCase() + option.slice(1)}</span>
-        </button>
-      {/each}
-    </div>
+    <ToggleSegmentedControl
+      {value}
+      onchange={(v) => { value = v; onchange({ toggleId: toggle.toggleId, value: v }); }}
+    />
   </div>
 </div>
 
@@ -74,63 +56,5 @@
     font-size: 0.75rem;
     color: var(--cv-text-secondary);
     margin: 0.125rem 0 0 0;
-  }
-
-  .segmented {
-    display: flex;
-    border: 1px solid var(--cv-border);
-    border-radius: 0.375rem;
-    overflow: hidden;
-    flex-shrink: 0;
-  }
-
-  .segment-btn {
-    display: flex;
-    align-items: center;
-    gap: 0.25rem;
-    background: transparent;
-    border: none;
-    border-left: 1px solid var(--cv-border);
-    padding: 0.3rem 0.6rem;
-    font-size: 0.8rem;
-    font-weight: 500;
-    color: var(--cv-text-secondary);
-    cursor: pointer;
-    transition: background 0.15s ease, color 0.15s ease;
-    font-family: inherit;
-    line-height: 1;
-  }
-
-  .segment-icon {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 14px;
-    height: 14px;
-    flex-shrink: 0;
-  }
-
-  .segment-icon :global(svg) {
-    width: 100%;
-    height: 100%;
-  }
-
-  .segment-label {
-    font-size: 0.75rem;
-  }
-
-  .segment-btn:first-child {
-    border-left: none;
-  }
-
-  .segment-btn:hover:not(.active) {
-    background: var(--cv-bg-hover);
-    color: var(--cv-text);
-  }
-
-  .segment-btn.active {
-    background: var(--cv-primary);
-    color: white;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.15);
   }
 </style>
